+++ src/components/Forum/ForumModal.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+
+const Modal: React.FC<{ onClose: () => void; children: React.ReactNode }> = ({
+  onClose,
+  children,
+}) => (
+  <div className="fixed inset-0 z-50 flex items-center justify-center bg-black bg-opacity-40">
+    <div className="bg-white rounded-lg shadow-lg p-6 max-w-2xl w-full relative">
+      <button
+        onClick={onClose}
+        className="absolute top-2 right-2 text-gray-500 hover:text-black text-xl"
+      >
+        &times;
+      </button>
+      {children}
+    </div>
+  </div>
+);
+
+export default Modal;
